// frontend/src/generated/PriceTable.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PriceEntry } from './PriceEntry'
import type { JsonValue } from './serde_json/JsonValue'

export type PriceTable = {
  schemaVersion: number
  entries: Array<PriceEntry>
  extra: { [key in string]?: JsonValue }
}